Fix/building docs (#3088)
* Fix building docs with babel 7

documentation.js required version 6 of babel, but phoenix has migrated to
version 7. To fix this, the dependency has been updated to
v9.0.0-alpha.0 which supports the latest babel.

Although this is an alpha dependency, it is specific to building the
docs, so as long as they build correctly, I don't see a huge issue with
it, it is preferable to downgrading babel.

* Update version in assets/package.json

This version is used when building the documentation. It has also been
added to the list of files with a version in RELEASE.md